match: reconstruct the memory allocator family
A bump arena with a 16-entry release stack, falling back to malloc when
the cursor is null. The allocation depth and entries are one u32[17]
object in both images: allocate and release index the entries from the
depth word through one base register, which the compiler emits only for a
single array. Six of the seven routines are exact; memory_allocate keeps
an unattributed register-assignment residue (size copied to a2 before the
cursor address takes a0). actor_try_attack_player is exact with the angle
held as a plain s32. GAME.EXE: 153 of 186 enrolled functions exact.

+++ src/game/actor.c
@@ -420,7 +420,7 @@ void actor_try_attack_player(
     KfActorDefinition *definition = actor_state.current_definition;
     KfActor *actor = actor_state.current;
     s32 distance;
-    u16 angle;
+    s32 angle;
     u16 status_effect;
 
     distance = actor_distance_to_point(
